+++ frontend/pages/entries/index.js
@@ -4,17 +4,20 @@ import SpendingContext from '../../componets/context';
 
 const Entries = () => {
   const deleteEntry = async (cookie, id) => {
-    // TODO abstract all this away
-    await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/entries/${id}`, {
-      method: 'DELETE',
-      credentials: 'include',
-      headers: {
-        cookie,
-        Accept: 'application/json',
-        'Content-Type': 'application/json'
-      }
-    });
-    location.reload();
+    const shouldDelete = window.confirm('Are you sure you want to delete this?');
+    if (shouldDelete) {
+      // TODO abstract all this away
+      await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/entries/${id}`, {
+        method: 'DELETE',
+        credentials: 'include',
+        headers: {
+          cookie,
+          Accept: 'application/json',
+          'Content-Type': 'application/json'
+        }
+      });
+      location.reload();
+    }
   };
 
   return (
@@ -25,7 +28,11 @@ const Entries = () => {
           {ctx.entries.map(c => {
             return (
               <div key={c._id}>
-                Name: {c.name}
+                <div> Name: {c.name}</div>
+                <div>Amount: {c.amount}</div>
+                <div>Category: {c.category?.name}</div>
+                <div>Method: {c.paymentMethod?.name}</div>
+                <div>Date: {new Date(c.date).toString()}</div>
                 <Link href={`/entries/${c._id}`}>
                   <a>Edit Entry</a>
                 </Link>
